requirements.txt
python-binance
bitmex-client
tensorflow
pygame
pandas
python-telegram-bot